Add dm2doxy doxygen filter
Works by parsing the whole environment and creating dummy files
corresponding to each real file, which are then served instead of
parsing the real files one-by-one.

Not recommended for use, as Doxygen takes upwards of 6.5 hours to
document /tg/station's 17,000 classes, completely independent of the
filter which finishes in little more than a minute.